Move & rename files to match project convention
