+++ src/components/sponsorBadge/index.js
