+++ src/pages/sponsors.js
@@ -2,8 +2,8 @@ import React, { Fragment } from 'react'
 import styled from 'styled-components'
 import { Container } from '../components/grid/'
 import ImageBanner from '../components/imageBanner'
-import SponsorBadge from '../templates/sponsors/SponsorBadge'
-import SponsorsSubsection from '../templates/sponsors/SponsorsSubsection'
+import SponsorBadge from '../components/sponsorBadge'
+import SponsorsSubsection from '../components/sponsorSubsection'
 import constants from '../constants'
 import theme from '../theme/theme'
 import { media } from '../theme/media'
